Keep just the two tests in v0.rs
Move Commands into to ../commands.

tested via:

BITCOIND_TEST=1 MADHOUSE=0 rm -rf /tmp/stacks-node-tests/ && clear &&
cargo test tests::signer::v0::\
allow_reorg_within_first_proposal_burn_block_timing_secs_madhouse \
-- --nocapture --test-threads=1 | \
awk '/running [1-9][0-9]* test/,/test result:/'

Checking: Shutting down miners. Result: true
Applying: Shutting down miners
Selected:
01. SKIP_COMMIT_OP_SECONDARY_MINER
02. BOOT_TO_EPOCH_3
03. SKIP_COMMIT_OP_PRIMARY_MINER
04. MINE_BITCOIN_BLOCK_AND_TENURE_CHANGE_MINER_1
05. SUBMIT_BLOCK_COMMIT_SECONDARY_MINER
06. STALL_MINING
07. MINE_TENURE
08. SUBMIT_BLOCK_COMMIT_PRIMARY_MINER
09. RECOVER_FROM_STALL
10. WAIT_FOR_BLOCK_FROM_MINER_2
11. MINE_TENURE
12. WAIT_FOR_BLOCK_FROM_MINER_1
13. SUBMIT_BLOCK_COMMIT_PRIMARY_MINER
14. SEND_TRANSFER_TX
15. WAIT_FOR_BLOCK_FROM_MINER_1
16. MINE_BITCOIN_BLOCK_AND_TENURE_CHANGE_MINER_1
17. SHUTDOWN_MINERS
Executed:
01. SKIP_COMMIT_OP_SECONDARY_MINER (1.00µs)
02. BOOT_TO_EPOCH_3 (108.05s)
03. SKIP_COMMIT_OP_PRIMARY_MINER (834.00ns)
04. MINE_BITCOIN_BLOCK_AND_TENURE_CHANGE_MINER_1 (2.10s)
05. SUBMIT_BLOCK_COMMIT_SECONDARY_MINER (1.00s)
06. STALL_MINING (833.00ns)
07. MINE_TENURE (1.00s)
08. SUBMIT_BLOCK_COMMIT_PRIMARY_MINER (501.37ms)
09. RECOVER_FROM_STALL (500.00ns)
10. WAIT_FOR_BLOCK_FROM_MINER_2 (500.74ms)
11. MINE_TENURE (1.00s)
12. WAIT_FOR_BLOCK_FROM_MINER_1 (1.00s)
13. SUBMIT_BLOCK_COMMIT_PRIMARY_MINER (501.32ms)
14. SEND_TRANSFER_TX (10.81ms)
15. WAIT_FOR_BLOCK_FROM_MINER_1 (5.01s)
16. MINE_BITCOIN_BLOCK_AND_TENURE_CHANGE_MINER_1 (1.56s)
17. SHUTDOWN_MINERS (2.75µs)

ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 245 filtered out;
finished in 145.35s
