+++ testnet/stacks-node/src/tests/signer/mod.rs
@@ -12,6 +12,7 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
+mod commands;
 mod v0;
 
 use std::collections::HashSet;
