+++ testnet/stacks-node/src/tests/signer/v0.rs
@@ -102,6 +102,20 @@ use crate::tests::{
 };
 use crate::{nakamoto_node, BitcoinRegtestController, BurnchainController, Config, Keychain};
 
+use crate::neon::RunLoopCounter;
+use crate::tests::signer::commands::*;
+use madhouse::{execute_commands, prop_allof, scenario, Command};
+use proptest::prelude::Strategy;
+use stacks::net::api::getinfo::RPCPeerInfoData;
+
+pub fn get_chain_info_wrapper(conf: &Config) -> RPCPeerInfoData {
+    get_chain_info(conf)
+}
+
+pub fn test_mine_stall_set(value: bool) {
+    TEST_MINE_STALL.set(value)
+}
+
 impl SignerTest<SpawnedSigner> {
     /// Run the test until the first epoch 2.5 reward cycle.
     /// Will activate pox-4 and register signers for the first full Epoch 2.5 reward cycle.
@@ -613,6 +627,32 @@ impl MultipleMinerTest {
         }
     }
 
+    pub fn get_primary_skip_commit_flag(&self) -> stacks::util::tests::TestFlag<bool> {
+        self.signer_test
+            .running_nodes
+            .counters
+            .naka_skip_commit_op
+            .clone()
+    }
+
+    pub fn get_secondary_skip_commit_flag(&self) -> stacks::util::tests::TestFlag<bool> {
+        self.rl2_counters.naka_skip_commit_op.clone()
+    }
+
+    pub fn get_primary_last_stacks_tip_counter(&self) -> RunLoopCounter {
+        self.signer_test
+            .running_nodes
+            .counters
+            .naka_submitted_commit_last_stacks_tip
+            .clone()
+    }
+
+    pub fn get_secondary_last_stacks_tip_counter(&self) -> RunLoopCounter {
+        self.rl2_counters
+            .naka_submitted_commit_last_stacks_tip
+            .clone()
+    }
+
     /// Boot node 1 to epoch 3.0 and wait for node 2 to catch up.
     pub fn boot_to_epoch_3(&mut self) {
         info!(
@@ -1039,7 +1079,7 @@ fn wait_for_block_pushed(
 }
 
 /// Waits for a block with the provided expected height to be proposed and pushed by the miner with the provided public key.
-fn wait_for_block_pushed_by_miner_key(
+pub fn wait_for_block_pushed_by_miner_key(
     timeout_secs: u64,
     expected_height: u64,
     miner_key: &StacksPublicKey,
@@ -12108,767 +12148,6 @@ fn mark_miner_as_invalid_if_reorg_is_rejected() {
         );
     }
     miners.shutdown();
-}
-
-use proptest::prelude::{Just, Strategy};
-use std::fmt::Debug;
-
-use madhouse::{
-    execute_commands, prop_allof, scenario, Command, CommandWrapper, State, TestContext,
-};
-
-pub struct SignerTestContext {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl Debug for SignerTestContext {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("SignerTestContext").finish()
-    }
-}
-
-impl Clone for SignerTestContext {
-    fn clone(&self) -> Self {
-        Self {
-            miners: self.miners.clone(),
-        }
-    }
-}
-
-impl TestContext for SignerTestContext {}
-
-impl SignerTestContext {
-    fn new(num_signers: usize, num_transfer_txs: u64) -> Self {
-        let miners = MultipleMinerTest::new_with_config_modifications(
-            num_signers,
-            num_transfer_txs,
-            |signer_config| {
-                signer_config.block_proposal_validation_timeout = Duration::from_secs(1800);
-                signer_config.tenure_last_block_proposal_timeout = Duration::from_secs(1800);
-                signer_config.first_proposal_burn_block_timing = Duration::from_secs(1800);
-            },
-            |config| {
-                config.miner.block_commit_delay = Duration::from_secs(0);
-            },
-            |config| {
-                config.miner.block_commit_delay = Duration::from_secs(0);
-            },
-        );
-
-        Self {
-            miners: Arc::new(Mutex::new(miners)),
-        }
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct SignerTestState {
-    is_booted_to_nakamoto: bool,
-    is_primary_miner_skip_commit_op: bool,
-    is_secondary_miner_skip_commit_op: bool,
-    mining_stalled: bool,
-}
-
-impl State for SignerTestState {}
-
-struct BootToEpoch3 {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl BootToEpoch3 {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for BootToEpoch3 {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Booting miners to Nakamoto. Result: {:?}",
-            !state.is_booted_to_nakamoto
-        );
-        !state.is_booted_to_nakamoto
-    }
-
-    fn apply(&self, state: &mut SignerTestState) {
-        println!("Applying: Booting miners to Nakamoto");
-
-        self.miners.lock().unwrap().boot_to_epoch_3();
-
-        let (conf_1, _) = self.miners.lock().unwrap().get_node_configs();
-        let burn_block_height = get_chain_info(&conf_1).burn_block_height;
-
-        assert_eq!(burn_block_height, 231);
-
-        state.is_booted_to_nakamoto = true;
-    }
-
-    fn label(&self) -> String {
-        "BOOT_TO_EPOCH_3".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(BootToEpoch3::new(ctx.miners.clone())))
-    }
-}
-
-struct SkipCommitOpPrimaryMiner {
-    miner_1_skip_commit_flag: stacks::util::tests::TestFlag<bool>,
-}
-
-impl SkipCommitOpPrimaryMiner {
-    pub fn new(miner_1_skip_commit_flag: stacks::util::tests::TestFlag<bool>) -> Self {
-        Self {
-            miner_1_skip_commit_flag,
-        }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for SkipCommitOpPrimaryMiner {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Skipping commit operations for miner 1. Result: {:?}",
-            !state.is_primary_miner_skip_commit_op
-        );
-        // Check if the miner has booted to Nakamoto.
-        !state.is_primary_miner_skip_commit_op
-    }
-
-    fn apply(&self, state: &mut SignerTestState) {
-        println!("Applying: Skipping commit operations for miner 1");
-
-        self.miner_1_skip_commit_flag.set(true);
-
-        state.is_primary_miner_skip_commit_op = true;
-    }
-
-    fn label(&self) -> String {
-        "SKIP_COMMIT_OP_PRIMARY_MINER".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(SkipCommitOpPrimaryMiner::new(
-            ctx.miners
-                .lock()
-                .unwrap()
-                .signer_test
-                .running_nodes
-                .counters
-                .naka_skip_commit_op
-                .clone(),
-        )))
-    }
-}
-
-struct SkipCommitOpSecondaryMiner {
-    miner_2_skip_commit_flag: stacks::util::tests::TestFlag<bool>,
-}
-
-impl SkipCommitOpSecondaryMiner {
-    pub fn new(miner_2_skip_commit_flag: stacks::util::tests::TestFlag<bool>) -> Self {
-        Self {
-            miner_2_skip_commit_flag,
-        }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for SkipCommitOpSecondaryMiner {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Skipping commit operations for miner 2. Result: {:?}",
-            !state.is_secondary_miner_skip_commit_op
-        );
-        // Check if the miner has not already skipped the commit operations.
-        !state.is_secondary_miner_skip_commit_op
-    }
-
-    fn apply(&self, state: &mut SignerTestState) {
-        println!("Applying: Skipping commit operations for miner 2");
-
-        self.miner_2_skip_commit_flag.set(true);
-
-        state.is_secondary_miner_skip_commit_op = true;
-    }
-
-    fn label(&self) -> String {
-        "SKIP_COMMIT_OP_SECONDARY_MINER".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(SkipCommitOpSecondaryMiner::new(
-            ctx.miners
-                .lock()
-                .unwrap()
-                .rl2_counters
-                .naka_skip_commit_op
-                .clone(),
-        )))
-    }
-}
-
-struct MineBitcoinBlockTenureChangePrimaryMinerCommand {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl MineBitcoinBlockTenureChangePrimaryMinerCommand {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext>
-    for MineBitcoinBlockTenureChangePrimaryMinerCommand
-{
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Miner 1 mining Bitcoin block and tenure change tx. Result: {:?}",
-            state.is_booted_to_nakamoto
-        );
-        // Check if the miner has booted to Nakamoto.
-        state.is_booted_to_nakamoto
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!("Applying: Miner 1 mining Bitcoin block and tenure change tx");
-
-        let (stacks_height_before, conf_1, miner_pk_1) = {
-            let mut miners = self.miners.lock().unwrap();
-            let stacks_height_before = miners.get_peer_stacks_tip_height();
-            let (conf_1, _) = miners.get_node_configs();
-            let burnchain = conf_1.get_burnchain();
-            let sortdb = burnchain.open_sortition_db(true).unwrap();
-
-            miners
-                .mine_bitcoin_block_and_tenure_change_tx(&sortdb, TenureChangeCause::BlockFound, 60)
-                .expect("Failed to mine BTC block");
-
-            let (miner_pk_1, _) = miners.get_miner_public_keys();
-
-            (stacks_height_before, conf_1, miner_pk_1)
-        };
-
-        println!(
-            "Waiting for Nakamoto block {} pushed by miner 1",
-            stacks_height_before + 1
-        );
-
-        let miner_1_block =
-            wait_for_block_pushed_by_miner_key(30, stacks_height_before + 1, &miner_pk_1)
-                .expect("Failed to get block");
-
-        let mined_block_height = miner_1_block.header.chain_length;
-        info!(
-            "Miner 1 mined Nakamoto block height: {}",
-            mined_block_height
-        );
-
-        let info_after = get_chain_info(&conf_1);
-        assert_eq!(info_after.stacks_tip, miner_1_block.header.block_hash());
-        assert_eq!(info_after.stacks_tip_height, mined_block_height);
-        assert_eq!(mined_block_height, stacks_height_before + 1);
-
-        // TODO: Verify Sortition winner based on state and context.
-    }
-
-    fn label(&self) -> String {
-        "MINE_BITCOIN_BLOCK_AND_TENURE_CHANGE_MINER_1".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(
-            MineBitcoinBlockTenureChangePrimaryMinerCommand::new(ctx.miners.clone()),
-        ))
-    }
-}
-
-struct MineBitcoinBlockTenureChangeSecondaryMinerCommand {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl MineBitcoinBlockTenureChangeSecondaryMinerCommand {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext>
-    for MineBitcoinBlockTenureChangeSecondaryMinerCommand
-{
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Miner 2 mining Bitcoin block and tenure change tx. Result: {:?}",
-            state.is_booted_to_nakamoto
-        );
-        // Check if the miner has booted to Nakamoto.
-        state.is_booted_to_nakamoto
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!("Applying: Miner 2 mining Bitcoin block and tenure change tx");
-
-        let stacks_height_before = self.miners.lock().unwrap().get_peer_stacks_tip_height();
-
-        let (conf_1, conf_2) = self.miners.lock().unwrap().get_node_configs();
-        let burnchain = conf_1.get_burnchain();
-        let sortdb = burnchain.open_sortition_db(true).unwrap();
-        self.miners
-            .lock()
-            .unwrap()
-            .mine_bitcoin_block_and_tenure_change_tx(&sortdb, TenureChangeCause::BlockFound, 60)
-            .expect("Failed to mine BTC block");
-
-        let (miner_pk_2, _) = self.miners.lock().unwrap().get_miner_public_keys();
-
-        println!(
-            "Waiting for Nakamoto block {} pushed by miner 2",
-            stacks_height_before + 1
-        );
-
-        let secondary_miner_block =
-            wait_for_block_pushed_by_miner_key(30, stacks_height_before + 1, &miner_pk_2)
-                .expect("Failed to get block N");
-
-        let mined_block_height = secondary_miner_block.header.chain_length;
-
-        let info_after = get_chain_info(&conf_2);
-        assert_eq!(
-            info_after.stacks_tip,
-            secondary_miner_block.header.block_hash()
-        );
-        assert_eq!(info_after.stacks_tip_height, mined_block_height);
-        assert_eq!(mined_block_height, stacks_height_before + 1);
-
-        // TODO: Verify Sortition winner based on state and context.
-    }
-
-    fn label(&self) -> String {
-        "MINE_BITCOIN_BLOCK_AND_TENURE_CHANGE_MINER_2".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(
-            MineBitcoinBlockTenureChangeSecondaryMinerCommand::new(ctx.miners.clone()),
-        ))
-    }
-}
-
-struct WaitForBlockFromMiner1Command {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl WaitForBlockFromMiner1Command {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for WaitForBlockFromMiner1Command {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Waiting for Nakamoto block from miner 1. Result: {:?}",
-            !state.mining_stalled
-        );
-        // Check if mining is no longer stalled
-        !state.mining_stalled
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!("Applying: Waiting for Nakamoto block from miner 1");
-
-        let miners_arc = self.miners.clone();
-
-        let (miner_pk_1, last_confirmed_nakamoto_height_counter) = {
-            let miners = miners_arc.lock().unwrap();
-            let (miner_pk_1, _) = miners.get_miner_public_keys();
-            let last_confirmed_nakamoto_height = miners
-                .signer_test
-                .running_nodes
-                .counters
-                .naka_submitted_commit_last_stacks_tip
-                .clone();
-            (miner_pk_1, last_confirmed_nakamoto_height)
-        };
-
-        let last_confirmed_height = last_confirmed_nakamoto_height_counter
-            .0
-            .load(std::sync::atomic::Ordering::SeqCst);
-        let expected_height = last_confirmed_height + 1;
-
-        println!(
-            "Waiting for Nakamoto block {} pushed by miner 1",
-            expected_height
-        );
-
-        let _miner_1_block = wait_for_block_pushed_by_miner_key(30, expected_height, &miner_pk_1)
-            .expect(&format!("Failed to get block {}", expected_height));
-    }
-
-    fn label(&self) -> String {
-        "WAIT_FOR_BLOCK_FROM_MINER_1".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(WaitForBlockFromMiner1Command::new(
-            ctx.miners.clone(),
-        )))
-    }
-}
-
-struct WaitForBlockFromMiner2Command {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl WaitForBlockFromMiner2Command {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for WaitForBlockFromMiner2Command {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Waiting for Nakamoto block from miner 2. Result: {:?}",
-            !state.mining_stalled
-        );
-        // Check if mining is no longer stalled
-        !state.mining_stalled
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!("Applying: Waiting for Nakamoto block from miner 2");
-
-        let miners_arc = self.miners.clone();
-
-        let (miner_pk_2, last_confirmed_nakamoto_height_counter) = {
-            let miners = miners_arc.lock().unwrap();
-            let (_, miner_pk_2) = miners.get_miner_public_keys();
-            let last_confirmed_nakamoto_height = miners
-                .rl2_counters
-                .naka_submitted_commit_last_stacks_tip
-                .clone();
-
-            (miner_pk_2, last_confirmed_nakamoto_height)
-        };
-
-        let last_confirmed_height = last_confirmed_nakamoto_height_counter
-            .0
-            .load(std::sync::atomic::Ordering::SeqCst);
-        let expected_stacks_height = last_confirmed_height + 1;
-
-        println!(
-            "Waiting for Nakamoto block {} pushed by miner 2",
-            expected_stacks_height
-        );
-
-        let _miner_2_block_n_1 =
-            wait_for_block_pushed_by_miner_key(30, expected_stacks_height, &miner_pk_2)
-                .expect(&format!("Failed to get block {:?}", expected_stacks_height));
-    }
-
-    fn label(&self) -> String {
-        "WAIT_FOR_BLOCK_FROM_MINER_2".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(WaitForBlockFromMiner2Command::new(
-            ctx.miners.clone(),
-        )))
-    }
-}
-
-struct SubmitBlockCommitSecondaryMinerCommand {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl SubmitBlockCommitSecondaryMinerCommand {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for SubmitBlockCommitSecondaryMinerCommand {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Submitting block commit miner 2. Result: {:?}",
-            state.is_secondary_miner_skip_commit_op
-        );
-        // Check if the secondary miner is skipping commit operations. If yes,
-        // we can manually submit the block commit. Otherwise, applying this
-        // may result in no commit being submitted.
-        state.is_secondary_miner_skip_commit_op
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!("Applying: Submitting block commit miner 2");
-
-        let (conf_1, _) = self.miners.lock().unwrap().get_node_configs();
-        let burnchain = conf_1.get_burnchain();
-        let sortdb = burnchain.open_sortition_db(true).unwrap();
-
-        self.miners.lock().unwrap().submit_commit_miner_2(&sortdb);
-    }
-
-    fn label(&self) -> String {
-        "SUBMIT_BLOCK_COMMIT_SECONDARY_MINER".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(
-            SubmitBlockCommitSecondaryMinerCommand::new(ctx.miners.clone()),
-        ))
-    }
-}
-
-struct SubmitBlockCommitPrimaryMinerCommand {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl SubmitBlockCommitPrimaryMinerCommand {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for SubmitBlockCommitPrimaryMinerCommand {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Submitting block commit miner 1. Result: {:?}",
-            state.is_primary_miner_skip_commit_op
-        );
-        // Check if the primary miner is skipping commit operations. If yes, we
-        // can manually submit the block commit. Otherwise, applying this may
-        // result in no commit being submitted.
-        state.is_primary_miner_skip_commit_op
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!("Applying: Submitting block commit miner 1");
-
-        let (conf_1, _) = self.miners.lock().unwrap().get_node_configs();
-        let burnchain = conf_1.get_burnchain();
-        let sortdb = burnchain.open_sortition_db(true).unwrap();
-
-        self.miners.lock().unwrap().submit_commit_miner_1(&sortdb);
-    }
-
-    fn label(&self) -> String {
-        "SUBMIT_BLOCK_COMMIT_PRIMARY_MINER".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(
-            SubmitBlockCommitPrimaryMinerCommand::new(ctx.miners.clone()),
-        ))
-    }
-}
-
-struct StallMiningCommand;
-
-impl Command<SignerTestState, SignerTestContext> for StallMiningCommand {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Stalling mining. Result: {:?}",
-            !state.mining_stalled
-        );
-        // Check if mining is not already stalled.
-        !state.mining_stalled
-    }
-
-    fn apply(&self, state: &mut SignerTestState) {
-        println!("Applying: Stalling mining");
-        TEST_MINE_STALL.set(true);
-
-        state.mining_stalled = true;
-    }
-
-    fn label(&self) -> String {
-        "STALL_MINING".to_string()
-    }
-
-    fn build(
-        _ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(StallMiningCommand))
-    }
-}
-
-struct RecoverFromStallCommand;
-
-impl Command<SignerTestState, SignerTestContext> for RecoverFromStallCommand {
-    fn check(&self, state: &SignerTestState) -> bool {
-        println!(
-            "Checking: Recovering from mining stall. Result: {:?}",
-            state.mining_stalled
-        );
-        // Check if mining is stalled.
-        state.mining_stalled
-    }
-
-    fn apply(&self, state: &mut SignerTestState) {
-        println!("Applying: Recovering from mining stall");
-        TEST_MINE_STALL.set(false);
-
-        state.mining_stalled = false;
-    }
-
-    fn label(&self) -> String {
-        "RECOVER_FROM_STALL".to_string()
-    }
-
-    fn build(
-        _ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(RecoverFromStallCommand))
-    }
-}
-
-struct MineTenureCommand {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-    timeout_secs: u64,
-}
-
-impl MineTenureCommand {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>, timeout_secs: u64) -> Self {
-        Self {
-            miners,
-            timeout_secs,
-        }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for MineTenureCommand {
-    fn check(&self, _state: &SignerTestState) -> bool {
-        println!("Checking: Mining tenure. Result: {:?}", true);
-        // TODO: Always appliable for now. Rewrite to mimic the original test
-        // if needed.
-        true
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!(
-            "Applying: Mining tenure and waiting for it for {:?} seconds",
-            self.timeout_secs
-        );
-
-        let sortdb = {
-            let miners = self.miners.lock().unwrap();
-            let (conf_1, _) = miners.get_node_configs();
-            let burnchain = conf_1.get_burnchain();
-            let sortdb = burnchain.open_sortition_db(true).unwrap();
-            sortdb
-        };
-
-        {
-            let mut miners = self.miners.lock().unwrap();
-            miners
-                .mine_bitcoin_blocks_and_confirm(&sortdb, 1, self.timeout_secs)
-                .expect("Failed to mine BTC block");
-        }
-    }
-
-    fn label(&self) -> String {
-        "MINE_TENURE".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        (60u64..90u64).prop_map(move |timeout_secs| {
-            CommandWrapper::new(MineTenureCommand::new(ctx.miners.clone(), timeout_secs))
-        })
-    }
-}
-
-struct SendTransferTxCommand {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl SendTransferTxCommand {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for SendTransferTxCommand {
-    fn check(&self, _state: &SignerTestState) -> bool {
-        println!("Checking: Sending transfer tx. Result: {:?}", true);
-        true
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!("Applying: Sending transfer tx");
-
-        self.miners.lock().unwrap().send_transfer_tx();
-    }
-
-    fn label(&self) -> String {
-        "SEND_TRANSFER_TX".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(SendTransferTxCommand::new(
-            ctx.miners.clone(),
-        )))
-    }
-}
-
-struct ShutdownMinersCommand {
-    miners: Arc<Mutex<MultipleMinerTest>>,
-}
-
-impl ShutdownMinersCommand {
-    pub fn new(miners: Arc<Mutex<MultipleMinerTest>>) -> Self {
-        Self { miners }
-    }
-}
-
-impl Command<SignerTestState, SignerTestContext> for ShutdownMinersCommand {
-    fn check(&self, _state: &SignerTestState) -> bool {
-        println!("Checking: Shutting down miners. Result: {:?}", true);
-        true
-    }
-
-    fn apply(&self, _state: &mut SignerTestState) {
-        println!("Applying: Shutting down miners");
-
-        if let Ok(miners_arc) = Arc::try_unwrap(self.miners.clone()) {
-            if let Ok(miners) = miners_arc.into_inner() {
-                miners.shutdown();
-            }
-        }
-    }
-
-    fn label(&self) -> String {
-        "SHUTDOWN_MINERS".to_string()
-    }
-
-    fn build(
-        ctx: Arc<SignerTestContext>,
-    ) -> impl Strategy<Value = CommandWrapper<SignerTestState, SignerTestContext>> {
-        Just(CommandWrapper::new(ShutdownMinersCommand::new(
-            ctx.miners.clone(),
-        )))
-    }
 }
 
 #[test]
